+++ fltk/src/browser.rs
@@ -121,8 +121,8 @@ impl FileBrowser {
     /// The following syntax is used for the pattern:
     /// `*` matches any sequence of 0 or more characters.
     /// `?` matches any single character.
-    /// `[set]` matches any character in the set. The set can contain any single characters, or a-z to represent a range. To match `]` or
-    /// they must be the first characters. To match `^` or `!` they must not be the first characters.
+    /// `[set]` matches any character in the set. The set can contain any single characters, or a-z to represent a range. 
+    /// To match `]` or `-`, they must be the first characters. To match `^` or `!`, they must not be the first characters.
     /// `[^set]` or `[!set]` matches any character not in the set.
     /// `{X|Y|Z}` or `{X,Y,Z}` matches any one of the subexpressions literally.
     /// `\x` quotes the character `x` so it has no special meaning.